re-organized component directories
