+++ components/detail/RoomDetail.vue
@@ -1,0 +1,8 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <div>
+    <h1>Room Detail</h1>
+  </div>
+</template>
